code enhancement

+++ src/components/profile/saga.jsx
@@ -39,9 +39,16 @@ function* changeEmailSaga(data) {
       yield put({ type: UPDATE_EMAIL_ERROR, payload: res.message });
     } else {
       yield put({ type: UPDATE_EMAIL_SUCCESS, payload: res.message });
-      yield put({ type: LOGOUT_SUCCESS });
-      localStorage.clear();
-      history.push("/");
+      if (
+        res &&
+        res.result &&
+        res.result._id &&
+        res.result._id.toString() === localStorage.getItem("userId").toString()
+      ) {
+        yield put({ type: LOGOUT_SUCCESS });
+        localStorage.clear();
+        history.push("/");
+      }
     }
   } catch (error) {
     yield put({ type: UPDATE_EMAIL_ERROR, payload: error.message });
@@ -56,9 +63,16 @@ function* resetPasswordSaga(data) {
       yield put({ type: UPDATE_PASSWORD_ERROR, payload: res.message });
     } else {
       yield put({ type: UPDATE_PASSWORD_SUCCESS, payload: res.message });
-      yield put({ type: LOGOUT_SUCCESS });
-      localStorage.clear();
-      history.push("/");
+      if (
+        res &&
+        res.result &&
+        res.result._id &&
+        res.result._id.toString() === localStorage.getItem("userId").toString()
+      ) {
+        yield put({ type: LOGOUT_SUCCESS });
+        localStorage.clear();
+        history.push("/");
+      }
     }
   } catch (error) {
     yield put({ type: UPDATE_PASSWORD_ERROR, payload: error.message });
